+++ CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes to this project will be documented in this file.
 ## [unreleased]
 ### Added
 - Added the solc version in truffle.js
+- Upgrade openzeppelin-solidity to 2.0.0
 
 ## [0.2.0] - 2018-11-28
 ### Added